fix(public_api): strengthen traversal tests and improve code organization
- Add ../../README, ../../AGENTS, ../../CODE_OF_CONDUCT to traversal payloads
  in unit and integration tests to ensure dict-lookup security property
- Fix _extract_title to strip fenced code blocks before searching for headings
- Move ConceptDocSlugConverter to converters.py and register in apps.py ready()
- Add explanatory comment to lookup_value_regex about OpenAPI schema generation
- Change authentication_classes from () to [] to match phase spec
